Update: 상준님 코드에 따른 수정 및 예외처리
