+++ pages/community/notice/detail/index.html
@@ -9,7 +9,7 @@
     <link rel="stylesheet" href="/components/footer/footer.css" />
     <script type="module" src="/components/component.js"></script>
     <link rel="stylesheet" href="style.css" />
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
     <title>Document</title>
   </head>
   <body>
